refactor(Theme): Split theme into different files
fix #23
